feature: shift from enum to enum class

+++ src/dam/keyboard_state.cpp
@@ -8,10 +8,10 @@ dam::input::KeyboardState::KeyboardState()
 
 bool dam::input::KeyboardState::is_key_down(dam::input::Key key)
 {
-    return al_key_down(state.get(), key);
+    return al_key_down(state.get(), static_cast<int>(key));
 }
 
 bool dam::input::KeyboardState::is_key_up(dam::input::Key key)
 {
-    return !al_key_down(state.get(), key);
+    return !al_key_down(state.get(), static_cast<int>(key));
 }
